pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer is a simple web application that checks the basic page-level elements of websites valuable for on-page SEO."
authors = ["Max <baiter159159@gmail.com>"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.3",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.9",
    "validators>=0.28.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "celery>=5.4.0",
    "redis>=5.0.4",
    "ruff>=0.9.2"
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "flake8>=7.1.1",
]